+++ integration/helpers/rsc-parcel/src/prerender.tsx
@@ -9,14 +9,14 @@ import { createFromReadableStream } from "react-server-dom-parcel/client.edge";
 
 export async function prerender(
   request: Request,
-  fetchServer: (request: Request) => Promise<Response>,
+  serverResponse: Response,
   bootstrapScriptContent: string | undefined,
 ): Promise<Response> {
   return await routeRSCServerRequest({
     // The incoming request.
     request,
     // How to fetch from the React Server.
-    fetchServer,
+    serverResponse,
     // Provide the React Server touchpoints.
     createFromReadableStream,
     // Render the router to HTML.