+++ integration/helpers/rsc-vite/src/entry.rsc.tsx
@@ -37,5 +37,5 @@ export default async function handler(request: Request) {
     // eslint-disable-next-line @typescript-eslint/consistent-type-imports
     typeof import("./entry.ssr")
   >("ssr", "index");
-  return ssr.default(request, fetchServer);
+  return ssr.default(request, await fetchServer(request));
 }