+++ playground/middleware/react-router.config.ts
@@ -3,6 +3,6 @@ import type { Config } from "@react-router/dev/config";
 export default {
   future: {
     v8_middleware: true,
-    unstable_splitRouteModules: true,
+    v8_splitRouteModules: true,
   },
 } satisfies Config;