+++ packages/create-react-router/CHANGELOG.md
@@ -1,8 +1,16 @@
 # `create-react-router`
 
+## 7.10.0
+
+_No changes_
+
 ## 7.9.6
 
+_No changes_
+
 ## 7.9.5
+
+_No changes_
 
 ## 7.9.4